@@ -1,10 +1,10 @@
 name: cardin_example
 description: Demonstrates Cardin theming
 publish_to: 'none'
-version: 1.0.0
+version: 0.1.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
