@@ -1,3 +1,4 @@
+import 'package:cardin/theme/colors.dart';
 import 'package:flutter/material.dart';
 import 'package:cardin/cardin.dart';
 
@@ -11,11 +12,12 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  var theme = cardinTheme;
+  ThemeData theme = cardinTheme;
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      debugShowCheckedModeBanner: false,
       title: 'Cardin Example',
       theme: theme,
       home: MyHomePage(
@@ -33,7 +35,7 @@ class _MyAppState extends State<MyApp> {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.themeChanged}) : super(key: key);
+  MyHomePage({Key? key, required this.themeChanged}) : super(key: key);
 
   final void Function(String themeName) themeChanged;
 
@@ -42,10 +44,7 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  final textController = TextEditingController(
-      text:
-          'My code fails, I do not know why.\nMy code works, I do not know why.\nText in other scripts: Tamaziɣt Taqbaylit, 中文(简体), Čeština, Беларуская, Ελληνικά, עברית, Русский, བོད་ཡིག, Norsk bokmål.');
-  var themeName = 'Cardin';
+  String themeName = 'Cardin';
   final void Function(String themeName) themeChanged;
 
   _MyHomePageState(this.themeChanged);
@@ -161,7 +160,7 @@ class _MyHomePageState extends State<MyHomePage> {
             Row(
               children: <Widget>[
                 DropdownButton<int>(
-                  onChanged: (value) => print('DropdownButton ${value}'),
+                  onChanged: (value) => print('DropdownButton $value'),
                   value: 1,
                   items: <DropdownMenuItem<int>>[
                     DropdownMenuItem(value: 1, child: Text('One')),
@@ -201,29 +200,6 @@ class _MyHomePageState extends State<MyHomePage> {
                   children: <Widget>[
                     Row(
                       children: <Widget>[
-                        Switch(onChanged: (value) {}, value: true),
-                        Text('Yes'),
-                      ],
-                    ),
-                    Row(
-                      children: <Widget>[
-                        Switch(onChanged: (value) {}, value: false),
-                        Text('No'),
-                      ],
-                    ),
-                    Row(
-                      children: <Widget>[
-                        Switch(onChanged: null, value: true),
-                        Text('Disabled'),
-                      ],
-                    ),
-                  ],
-                ),
-                Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: <Widget>[
-                    Row(
-                      children: <Widget>[
                         Radio(onChanged: (value) {}, value: 1, groupValue: 1),
                         Text('Yes'),
                       ],
@@ -237,6 +213,29 @@ class _MyHomePageState extends State<MyHomePage> {
                     Row(
                       children: <Widget>[
                         Radio(onChanged: null, value: 3, groupValue: 1),
+                        Text('Disabled'),
+                      ],
+                    ),
+                  ],
+                ),
+                Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: <Widget>[
+                    Row(
+                      children: <Widget>[
+                        Switch(onChanged: (value) {}, value: true),
+                        Text('Yes'),
+                      ],
+                    ),
+                    Row(
+                      children: <Widget>[
+                        Switch(onChanged: (value) {}, value: false),
+                        Text('No'),
+                      ],
+                    ),
+                    Row(
+                      children: <Widget>[
+                        Switch(onChanged: null, value: true),
                         Text('Disabled'),
                       ],
                     ),
@@ -267,104 +266,22 @@ class _MyHomePageState extends State<MyHomePage> {
               ),
             ),
             SizedBox(height: 15),
-            TextField(
-              keyboardType: TextInputType.multiline,
-              controller: textController,
-              minLines: 5,
-              maxLines: 5,
-              decoration: InputDecoration(
-                border: OutlineInputBorder(),
-              ),
-            ),
-            SizedBox(height: 15),
-            Container(
-              width: 250.0,
-              height: 50.0,
-              color: cardinBlue,
-              child: Center(
-                  child: Text('cardinBlue',
-                      style: TextStyle(
-                          color: cardinBlack,
-                          fontSize: 22,
-                          fontWeight: FontWeight.w400))),
-            ),
-            Container(
-              width: 250.0,
-              height: 50.0,
-              color: cardinRed,
-              child: Center(
-                  child: Text('cardinRed',
-                      style: TextStyle(
-                          color: cardinWhite,
-                          fontSize: 22,
-                          fontWeight: FontWeight.w400))),
-            ),
-            Container(
-              width: 250.0,
-              height: 50.0,
-              color: cardinGreen,
-              child: Center(
-                  child: Text('cardinGreen',
-                      style: TextStyle(
-                          color: cardinBlack,
-                          fontSize: 22,
-                          fontWeight: FontWeight.w400))),
-            ),
-            Container(
-              width: 250.0,
-              height: 50.0,
-              color: cardinMagenta,
-              child: Center(
-                  child: Text('cardinMagenta',
-                      style: TextStyle(
-                          color: cardinWhite,
-                          fontSize: 22,
-                          fontWeight: FontWeight.w400))),
-            ),
-            Container(
-              width: 250.0,
-              height: 50.0,
-              color: cardinCyan,
-              child: Center(
-                  child: Text('cardinCyan',
-                      style: TextStyle(
-                          color: cardinWhite,
-                          fontSize: 22,
-                          fontWeight: FontWeight.w400))),
-            ),
-            Container(
-              width: 250.0,
-              height: 50.0,
-              color: cardinYellow,
-              child: Center(
-                  child: Text('cardinYellow',
-                      style: TextStyle(
-                          color: cardinBlack,
-                          fontSize: 22,
-                          fontWeight: FontWeight.w400))),
-            ),
-            Container(
-              width: 250.0,
-              height: 50.0,
-              color: cardinWhite,
-              child: Center(
-                  child: Text('cardinWhite',
-                      style: TextStyle(
-                          color: cardinBlack,
-                          fontSize: 22,
-                          fontWeight: FontWeight.w400))),
-            ),
-            Container(
-              width: 250.0,
-              height: 50.0,
-              color: cardinBlack,
-              child: Center(
-                  child: Text('cardinBlack',
-                      style: TextStyle(
-                          color: cardinWhite,
-                          fontSize: 22,
-                          fontWeight: FontWeight.w400))),
-            ),
+            _getColowRow(cardinBlue, 'cardinBlue', cardinBrightBlue,
+                'cardinBrightBlue', cardinWhite),
+            _getColowRow(cardinRed, 'cardinRed', cardinBrightRed,
+                'cardinBrightRed', cardinWhite),
+            _getColowRow(cardinGreen, 'cardinGreen', cardinBrightGreen,
+                'cardinBrightGreeen', cardinWhite),
+            _getColowRow(cardinMagenta, 'cardinMagenta', cardinBrightMagenta,
+                'cardinBrightMagenta', cardinWhite),
+            _getColowRow(cardinCyan, 'cardinCyan', cardinBrightCyan,
+                'cardinBrightCyan', cardinWhite),
+            _getColowRow(cardinYellow, 'cardinYellow', cardinBrightYellow,
+                'cardinBrightYellow', cardinBlack),
+            _getColowRow(cardinWhite, 'cardinWhite', cardinBrightWhite,
+                'cardinBrightWhite', cardinBlack),
+            _getColowRow(cardinBlack, 'cardinBlack', cardinBrightBlack,
+                'cardinBrightBlack', cardinWhite),
           ],
         ),
       ),
@@ -374,5 +291,35 @@ class _MyHomePageState extends State<MyHomePage> {
         child: Icon(Icons.add),
       ),
     );
+  }
+
+  Row _getColowRow(Color color, String colorName, Color brightColor,
+      String brightColorName, Color textColor) {
+    return Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: <Widget>[
+          Expanded(
+              child: Container(
+            color: color,
+            height: 50,
+            child: Center(
+                child: Text(colorName,
+                    style: TextStyle(
+                        color: textColor,
+                        fontSize: 22,
+                        fontWeight: FontWeight.w400))),
+          )),
+          Expanded(
+              child: Container(
+            color: brightColor,
+            height: 50,
+            child: Center(
+                child: Text(brightColorName,
+                    style: TextStyle(
+                        color: textColor,
+                        fontSize: 22,
+                        fontWeight: FontWeight.w400))),
+          )),
+        ]);
   }
 }
